LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-replace-same-document-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected 3 but got 4

PASS navigate() with history: 'replace' option
